+++ CMakeLists.txt
@@ -711,6 +711,7 @@ if(WITH_OPTIM)
                 list(APPEND AVX512_SRCS ${ARCHDIR}/adler32_avx512.c)
                 add_feature_info(AVX512_ADLER32 1 "Support AVX512-accelerated adler32, using \"${AVX512FLAG}\"")
                 list(APPEND ZLIB_ARCH_SRCS ${AVX512_SRCS})
+                list(APPEND ZLIB_ARCH_HDRS ${ARCHDIR}/adler32_avx512_p.h)
                 if(HAVE_MASK_INTRIN)
                     add_definitions(-DX86_MASK_INTRIN)
                 endif()
